Upgrade Camel to 4.2.0
