+++ docs/antora.yml
@@ -28,7 +28,7 @@ asciidoc:
     requires: "'util=camel-website-util,quarkus=xref:js/quarkus.js'"
 
     # Project versions
-    camel-version: 4.1.0 # replace ${camel.version}
+    camel-version: 4.2.0 # replace ${camel.version}
     camel-docs-version: 4.0.x
     camel-quarkus-version: 3.6.0 # replace ${camel-quarkus.version}
     quarkus-version: 3.5.1 # replace ${quarkus.version}
