--- extensions/debug/deployment/src/test/java/org/apache/camel/quarkus/component/debug/deployment/DebugEnabledFromCamelMainPropertyTest.java
@@ -1,71 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.component.debug.deployment;
-
-import java.io.IOException;
-import java.io.StringWriter;
-import java.io.Writer;
-import java.net.Socket;
-import java.util.Properties;
-
-import io.quarkus.test.QuarkusUnitTest;
-import io.restassured.RestAssured;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.Asset;
-import org.jboss.shrinkwrap.api.asset.StringAsset;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import static org.hamcrest.Matchers.is;
-
-public class DebugEnabledFromCamelMainPropertyTest {
-
-    @RegisterExtension
-    static final QuarkusUnitTest CONFIG = new QuarkusUnitTest()
-            .setArchiveProducer(() -> ShrinkWrap.create(JavaArchive.class)
-                    .addClass(DebugResource.class)
-                    .addAsResource(applicationProperties(), "application.properties"));
-
-    @Test
-    public void camelDebuggingEnabledFromCamelMainProperty() throws IOException {
-        RestAssured.get("/debug/enabled")
-                .then()
-                .body(is("true"))
-                .statusCode(200);
-
-        try (Socket socket = new Socket("localhost", 1099)) {
-            Assertions.assertTrue(socket.isConnected());
-        }
-    }
-
-    public static final Asset applicationProperties() {
-        Writer writer = new StringWriter();
-
-        Properties props = new Properties();
-        props.setProperty("camel.main.debugging", "true");
-
-        try {
-            props.store(writer, "");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        return new StringAsset(writer.toString());
-    }
-}
