+++ integration-test-groups/foundation/type-converter/src/test/java/org/apache/camel/quarkus/core/converter/it/ConverterProfile1IT.java
@@ -14,27 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.quarkus.component.debug.deployment;
+package org.apache.camel.quarkus.core.converter.it;
 
-import jakarta.inject.Inject;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-import org.apache.camel.CamelContext;
-import org.apache.camel.impl.debugger.DefaultBacklogDebugger;
-import org.apache.camel.spi.BacklogDebugger;
+import io.quarkus.test.junit.QuarkusIntegrationTest;
 
-@Path("/debug")
-public class DebugResource {
-    @Inject
-    CamelContext context;
-
-    @Path("/enabled")
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public boolean debuggingEnabled() {
-        BacklogDebugger debugger = DefaultBacklogDebugger.getBacklogDebugger(context);
-        return debugger != null && debugger.isEnabled();
-    }
+@QuarkusIntegrationTest
+public class ConverterProfile1IT extends ConverterProfile1Test {
 }
