+++ integration-tests/camel-k-runtime/pom.xml
@@ -121,6 +121,7 @@
                 <quarkus.native.native-image-xmx>5g</quarkus.native.native-image-xmx>
             </properties>
         </profile>
+        <!-- disabled because of https://github.com/apache/camel-quarkus/issues/5447
         <profile>
             <id>native</id>
             <activation>
@@ -154,6 +155,7 @@
                 </plugins>
             </build>
         </profile>
+        -->
         <profile>
             <id>virtualDependencies</id>
             <activation>
