+++ integration-tests/kubernetes/pom.xml
@@ -81,6 +81,7 @@
                 <quarkus.native.native-image-xmx>5g</quarkus.native.native-image-xmx>
             </properties>
         </profile>
+        <!-- disabled because of https://github.com/apache/camel-quarkus/issues/5447
         <profile>
             <id>native</id>
             <activation>
@@ -108,6 +109,7 @@
                 </plugins>
             </build>
         </profile>
+        -->
         <profile>
             <id>virtualDependencies</id>
             <activation>
