+++ integration-tests/master-openshift/pom.xml
@@ -154,6 +154,7 @@
                 </plugins>
             </build>
         </profile>
+        <!-- disabled because of https://github.com/apache/camel-quarkus/issues/5447
          <profile>
              <id>native</id>
              <activation>
@@ -191,6 +192,7 @@
                  </plugins>
              </build>
          </profile>
+        -->
          <profile>
              <id>virtualDependencies</id>
              <activation>
